-Bump bnpl sdk version. -Remove unused attributes on bnpl schema. -Move providerLabel(languageTag:String!): String! from CreditLines to CreditLineProvider as poweredByLabel

+++ api/src/main/groovy/bff/model/BnplCreditLineQuery.groovy
@@ -20,7 +20,7 @@ class BnplCreditLineQuery implements GraphQLQueryResolver {
 
     CompletableFuture<CreditLinesResult> getCreditLines(CreditLinesRequestInput input) {
         def userId = JwtToken.userIdFromToken(input.getAccessToken())
-        bnPlSdk.fetchBalance(userId, input.getAccessToken())
+        bnPlSdk.fetchBalance(userId.toLong(), input.getAccessToken())
                 .map{
                     fromSdk(it)
                 }
